Remove ts-ignore/eslint comments, break out component from IconSearch
