+++ docs/types/evergreen-export.ts
@@ -1,0 +1,3 @@
+import * as Evergreen from 'evergreen-ui'
+
+export type EvergreenExport = keyof typeof Evergreen
